+++ src/components/nav-list/index.ts
@@ -1,0 +1,2 @@
+export * from "./NavList.js";
+export { NavList } from "./NavList.js";